requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
celery==5.3.1  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
SQLAlchemy==2.0.21
PyAMQP==0.1.0.7
redis==5.0.1
qrcode==7.4.2
django-jazzmin==2.6.0
sentry-sdk==1.31.0